+++ lib/Common/Memory/ArenaAllocator.h
@@ -817,14 +817,14 @@ public:
     {
     }
 
-    ulong AddRef(void)
+    uint32 AddRef(void)
     {
-        return (ulong)InterlockedIncrement(&refCount);
+        return (uint32)InterlockedIncrement(&refCount);
     }
 
-    ulong Release(void)
+    uint32 Release(void)
     {
-        ulong refs = (ulong)InterlockedDecrement(&refCount);
+        uint32 refs = (uint32)InterlockedDecrement(&refCount);
 
         if (0 == refs)
         {
@@ -882,7 +882,7 @@ public:
 class ReferencedArenaAdapter : public RefCounted
 {
     CRITICAL_SECTION adapterLock;
-    ulong strongRefCount;
+    uint32 strongRefCount;
     ArenaAllocator* arena;
     bool deleteFlag;
