+++ lib/Runtime/Debug/DiagObjectModel.h
@@ -53,7 +53,7 @@ namespace Js
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) = 0;
 
         // Returns number of children for the current diag object.
-        virtual ulong GetChildrenCount() = 0;
+        virtual uint32 GetChildrenCount() = 0;
 
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) = 0;
 
@@ -182,7 +182,7 @@ namespace Js
         /// IDiagObjectModelWalkerBase
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override sealed;
 
         virtual IDiagObjectAddress *FindPropertyAddress(PropertyId propId, bool& isConst) override;
@@ -274,7 +274,7 @@ namespace Js
     public:
         // Represent catch/with scope objects, (ie. the representation for the diagnostics purposes.)
         JsUtil::List<IDiagObjectModelWalkerBase*, ArenaAllocator> *pDiagScopeObjects;
-        ulong diagScopeVarCount;
+        uint32 diagScopeVarCount;
 
         bool scopeIsInitialized;
         bool enumWithScopeAlso;
@@ -287,7 +287,7 @@ namespace Js
         DiagScopeVariablesWalker(DiagStackFrame* _pFrame, Var _instance, IDiagObjectModelWalkerBase* innerWalker);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
 
         virtual IDiagObjectAddress *FindPropertyAddress(PropertyId propId, bool& isConst) override;
     };
@@ -311,7 +311,7 @@ namespace Js
     public:
         LocalsWalker(DiagStackFrame* _frame, DWORD _frameWalkerFlags);
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
 
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) {return FALSE; }
 
@@ -329,11 +329,11 @@ namespace Js
         {
             Assert(scriptContext);
             Js::DynamicObject* activeScopeObject = nullptr;
-            ulong count = this->GetChildrenCount();
+            uint32 count = this->GetChildrenCount();
             if (count > 0)
             {
                 activeScopeObject = scriptContext->GetLibrary()->CreateActivationObject();
-                for (ulong i = 0; i < count; i++)
+                for (uint32 i = 0; i < count; i++)
                 {
                     Js::ResolvedObject resolveObject;
                     if (this->Get(i, &resolveObject) && resolveObject.propId != Js::Constants::NoProperty)
@@ -441,7 +441,7 @@ namespace Js
         /// IDiagObjectModelWalkerBase
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override { return FALSE; }
         virtual IDiagObjectAddress *FindPropertyAddress(PropertyId propId, bool& isConst) override;
     private:
@@ -472,7 +472,7 @@ namespace Js
         RecyclableObjectWalker(ScriptContext* pContext, Var slot, Var originalInstance);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
 
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject)  { return FALSE; };
         virtual IDiagObjectAddress *FindPropertyAddress(PropertyId propertyId, bool& isConst) override;
@@ -579,7 +579,7 @@ namespace Js
         void SetOnlyWalkOwnProperties(bool set) { fOnlyOwnProperties = set; }
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
         virtual BOOL FetchItemAtIndex(Js::JavascriptArray* arrayObj, uint32 index, Var *value);
         virtual Var FetchItemAt(Js::JavascriptArray* arrayObj, uint32 index);
         virtual BOOL GetResolvedObject(Js::JavascriptArray* arrayObj, int index, ResolvedObject* pResolvedObject, uint32 * pabsIndex) sealed;
@@ -606,7 +606,7 @@ namespace Js
     public:
         RecyclableArgumentsObjectWalker(ScriptContext* pContext, Var instance, LocalsWalker * localsWalker);
 
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
     };
 
     class RecyclableArgumentsArrayAddress : public IDiagObjectAddress
@@ -627,7 +627,7 @@ namespace Js
         RecyclableArgumentsArrayWalker(ScriptContext* pContext, Var slot, Var originalInstance);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
 
         void FetchFormalsAddress (LocalsWalker * localsWalker);
     };
@@ -657,7 +657,7 @@ namespace Js
         RecyclableTypedArrayWalker(ScriptContext* pContext, Var slot, Var originalInstance);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
     };
 
     // Concrete classes for Pixel array objects
@@ -685,7 +685,7 @@ namespace Js
         RecyclablePixelArrayWalker(ScriptContext* pContext, Var slot, Var originalInstance);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
     };
 
     // Concrete classes for ES5 array objects
@@ -780,7 +780,7 @@ namespace Js
         RecyclableCollectionObjectWalker(ScriptContext* scriptContext, Var instance):scriptContext(scriptContext), instance(instance), propertyList(nullptr) { }
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override;
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
     };
 
     typedef RecyclableCollectionObjectWalker<JavascriptMap> RecyclableMapObjectWalker;
@@ -843,7 +843,7 @@ namespace Js
         RecyclableKeyValueWalker(ScriptContext* scriptContext, Var key, Var value):scriptContext(scriptContext), key(key), value(value) { }
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override { return 2; }
+        virtual uint32 GetChildrenCount() override { return 2; }
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override { return FALSE; }
     };
 
@@ -862,7 +862,7 @@ namespace Js
         RecyclableProxyObjectWalker(ScriptContext* pContext, Var instance);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override { return 2; }
+        virtual uint32 GetChildrenCount() override { return 2; }
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override;
     };
 
@@ -876,7 +876,7 @@ namespace Js
         void AddItem(Js::PropertyId propertyId, Var obj);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
         virtual BOOL GetGroupObject(ResolvedObject* pResolvedObject) override;
 
         void Sort();
@@ -951,7 +951,7 @@ namespace Js
         PendingMutationBreakpointWalker(ScriptContext* pContext, Var instance, MutationType mutationType);
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
-        virtual ulong GetChildrenCount() override;
+        virtual uint32 GetChildrenCount() override;
     };
 #endif
 
@@ -964,7 +964,7 @@ namespace Js
 
         virtual BOOL Get(int i, ResolvedObject* pResolvedObject) override;
 
-        virtual ulong GetChildrenCount() override { return elementCount; }
+        virtual uint32 GetChildrenCount() override { return elementCount; }
     };
 
     // For SIMD concrete walker: specify SIMD type and total elementCount
